feat: remove root span concept
Once the root span is being tracked on the receiver side it's not
possible for the telemetry producer to know when creating a nested span
whether there is a parent or not. So as preparation for that change make
the current implementation behave similarly with any span created while
there is no currently active trace act like a new root.

Signed-off-by: Wim Looman <wim.looman@veecle.io>