refactor(telemetry)!: remove explicit root spans (#89)
To support embedded devices we need to remove all thread-local state
from the telemetry library, this means we cannot know what the "current"
span is and have to move tracking of this to the consumer of the
telemetry messages. One feature of the current span tracking is to know
whether there is an active "root" span or not, only if there is a root
span do child spans actually get recorded, otherwise they're just
discarded when created. Without being able to track this in thread local
state we have to just assume that any span might be within an active
trace and must then record it.

This PR removes the root span APIs and changes the existing
implementation to act similar to how the final implementation will act,
a root span is just any span created while there is no "current span" on
this thread of execution.

Refs: DEV-911

---------

Signed-off-by: Wim Looman <wim.looman@veecle.io>