feat(orchestrator): implement control messages for runtime management (#91)
* Add support for marking runtimes as privileged via a new
`--privileged` flag in the orchestrator CLI. Privileged runtimes will be
allowed to send control messages through the IPC protocol.
Non-privileged runtimes attempting to send control messages will be
rejected with an error.

* Add `StartRuntime`, `StopRuntime`, and `GetStatus` control messages
that allow privileged runtimes to control other runtime instances.

* Add an actor that connects to the IPC channel and proxies
`ControlMessage` and `ControlResponse` messages back and forth between
IPC and the datastore.

* A simple example of a runtime that shuts itself down, both working and
failing because it was not registered privileged.

Closes: DEV-1019

---------

Signed-off-by: Wim Looman <wim.looman@veecle.io>