docs: add example to readme
The example showcases target agnostic actors leveraging the OSAL.
Logs are printed via the pretty print exporter.
Uses a git dependency to be able to use the pretty print exporter.
The `readme-example` crate checks the readme example in CI.

Fixes the license link in the readme to use the existing reference.

Fixes: DEV-1176

Signed-off-by: Julian Schindel <julian.schindel@veecle.io>
Co-authored-by: Wim Looman <wim.looman@veecle.io>