feat(telemetry-ui): Gray out filtered spans instead of hiding them (#145)
Instead of completely hiding spans that don't match active filters, they
are now displayed but are visuall grayed out.

Matching spans that are matched by the filter maintain 40% of the
grayness, but filtered-out spans are being dimed out (15% gray) with
reduced opacity to indicate that they are filtered out.

Demo:


https://github.com/user-attachments/assets/450c006d-9103-46e5-a538-b9a8951ecb76

Resolves: https://github.com/veecle/veecle-os/issues/138

Signed-off-by: tetektoza <tetektoza@gmail.com>